Move NMEA parsing into the parser module
